Cleaned up a lot of code.
