+++ src/index/CMakeLists.txt
@@ -6,4 +6,4 @@ add_library(index
         DocsDB.cpp FTSAlgorithms.cpp
         PrefixHeuristic.cpp CompressedRelation.cpp
         PatternCreator.cpp)
-qlever_target_link_libraries(index util parser vocabulary compilationInfo ${STXXL_LIBRARIES})
+qlever_target_link_libraries(index util parser vocabulary compilationInfo scanSpecification ${STXXL_LIBRARIES})
